Move relevant tests from hyperion to mx-bluesky
