+++ src/mx_bluesky/common/external_interaction/callbacks/common/abstract_event.py
