+++ src/mx_bluesky/common/parameters/gridscan.py
@@ -28,7 +28,6 @@ from mx_bluesky.common.parameters.constants import (
     HardwareConstants,
 )
 from mx_bluesky.common.parameters.robot_load import RobotLoadAndEnergyChange
-from mx_bluesky.hyperion.parameters.constants import CONST
 
 
 class GridCommon(
@@ -111,11 +110,11 @@ class ThreeDGridScan(
     """Parameters representing a so-called 3D grid scan, which consists of doing a
     gridscan in X and Y, followed by one in X and Z."""
 
-    grid1_omega_deg: float = Field(default=CONST.PARAM.GRIDSCAN.OMEGA_1)  # type: ignore
-    grid2_omega_deg: float = Field(default=CONST.PARAM.GRIDSCAN.OMEGA_2)
-    x_step_size_um: float = Field(default=CONST.PARAM.GRIDSCAN.BOX_WIDTH_UM)
-    y_step_size_um: float = Field(default=CONST.PARAM.GRIDSCAN.BOX_WIDTH_UM)
-    z_step_size_um: float = Field(default=CONST.PARAM.GRIDSCAN.BOX_WIDTH_UM)
+    grid1_omega_deg: float = Field(default=GridscanParamConstants.OMEGA_1)  # type: ignore
+    grid2_omega_deg: float = Field(default=GridscanParamConstants.OMEGA_2)
+    x_step_size_um: float = Field(default=GridscanParamConstants.BOX_WIDTH_UM)
+    y_step_size_um: float = Field(default=GridscanParamConstants.BOX_WIDTH_UM)
+    z_step_size_um: float = Field(default=GridscanParamConstants.BOX_WIDTH_UM)
     y2_start_um: float
     z2_start_um: float
     x_steps: int = Field(gt=0)
