+++ src/mx_bluesky/hyperion/external_interaction/callbacks/sample_handling/sample_handling_callback.py
@@ -8,6 +8,9 @@ from bluesky.preprocessors import contingency_wrapper
 from bluesky.utils import Msg, make_decorator
 from event_model import Event, EventDescriptor, RunStart
 
+from mx_bluesky.common.external_interaction.callbacks.common.abstract_event import (
+    AbstractEvent,
+)
 from mx_bluesky.common.external_interaction.callbacks.common.plan_reactive_callback import (
     PlanReactiveCallback,
 )
@@ -17,9 +20,6 @@ from mx_bluesky.common.external_interaction.ispyb.exp_eye_store import (
 )
 from mx_bluesky.common.utils.exceptions import CrystalNotFoundException, SampleException
 from mx_bluesky.common.utils.log import ISPYB_ZOCALO_CALLBACK_LOGGER
-from mx_bluesky.hyperion.external_interaction.callbacks.common.abstract_event import (
-    AbstractEvent,
-)
 from mx_bluesky.hyperion.parameters.constants import CONST
 
 # TODO remove this event-raising shenanigans once
