+++ tests/unit_tests/common/external_interaction/callbacks/common/test_abstract_event.py
@@ -4,7 +4,7 @@ import bluesky.plan_stubs as bps
 import bluesky.preprocessors as bpp
 from bluesky.run_engine import RunEngine
 
-from mx_bluesky.hyperion.external_interaction.callbacks.common.abstract_event import (
+from mx_bluesky.common.external_interaction.callbacks.common.abstract_event import (
     AbstractEvent,
 )
