+++ tests/unit_tests/hyperion/experiment_plans/test_multi_rotation_scan_plan.py
@@ -36,10 +36,10 @@ from ....conftest import (
     DocumentCapturer,
     extract_metafile,
     fake_read,
+    mx_acquisition_from_conn,
     raw_params_from_file,
 )
 from ..external_interaction.conftest import *  # noqa # for fixtures
-from ..external_interaction.conftest import mx_acquisition_from_conn
 
 TEST_OFFSET = 1
 TEST_SHUTTER_OPENING_DEGREES = 2.5
